# api/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRUD para Coder House",
        "version": "1.0.0",
        "title": "CRUD Tareas",
        "contact": {
            "name": "Franco Gómez Lisardia",
            "email": "lisardia.franco@gmail.com"
        }
    },
    "schemes": [
        "http"
    ],
    "paths": {
        "/tarea": {
            "post": {
                "tags": [
                    "tarea"
                ],
                "summary": "Agrega una nueva tarea a la lista",
                "description": "    ",
                "operationId": "addtarea",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto para agregar un item",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tarea"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Entrada inválida"
                    }
                },
                "security": [
                    {
                        "tareastore_auth": [
                            "write:tareas",
                            "read:tareas"
                        ]
                    }
                ]
            },
            "get": {
                "tags": [
                    "tarea"
                ],
                "summary": "Obtiene todas las tareas de la lista",
                "description": "    ",
                "operationId": "gettarea",
                "consumes": [
                    "application/json",
                    "application/xml"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "405": {
                        "description": "Entrada inválida"
                    }
                },
                "security": [
                    {
                        "tareastore_auth": [
                            "write:tareas",
                            "read:tareas"
                        ]
                    }
                ]
            }
        },
        "/tarea/{tareaId}": {
            "get": {
                "tags": [
                    "tarea"
                ],
                "summary": "Busca tarea por ID",
                "description": "Retorna una tarea",
                "operationId": "gettareaById",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "tareaId",
                        "in": "path",
                        "description": "ID de la tarea a retornar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/tarea"
                        }
                    },
                    "400": {
                        "description": "Id invalido"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "tarea"
                ],
                "summary": "Actualiza una tarea de la lista",
                "description": "",
                "operationId": "updatetareaWithForm",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "tareaId",
                        "in": "path",
                        "description": "ID de la terea a actualizar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "titulo",
                        "in": "formData",
                        "description": "Actualiza titulo de la tarea",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "descripcion",
                        "in": "formData",
                        "description": "Actualiza descripcion de la tarea",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Entrada invalida"
                    }
                },
                "security": [
                    {
                        "tareastore_auth": [
                            "write:tareas",
                            "read:tareas"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "tarea"
                ],
                "summary": "Elimina una tarea",
                "description": "",
                "operationId": "deletetarea",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "tareaId",
                        "in": "path",
                        "description": "ID de la tarea a eliminar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "ID invalido"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                },
                "security": [
                    {
                        "tareastore_auth": [
                            "write:tareas",
                            "read:tareas"
                        ]
                    }
                ]
            }
        }
    },
    "definitions": {
        "tarea": {
            "type": "object",
            "properties": {
                "titulo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "tarea"
            }
        }
    }
}